+++ web3/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @direct.dev/web3
 
+## 0.5.1
+
+### Patch Changes
+
+- Updated dependencies [28f2549]
+- Updated dependencies [36019eb]
+  - @direct.dev/client@0.5.1
+
 ## 0.5.0
 
 ### Patch Changes