+++ wire/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @direct.dev/wire
 
+## 0.5.1
+
+### Patch Changes
+
+- @direct.dev/shared@0.5.1
+
 ## 0.5.0
 
 ### Minor Changes